+++ src/Options.h
@@ -4,6 +4,5 @@
 #include "Error.h"
 #include "plugin/Plugin.h"
 
-
-Array Options_parse(
+Array* Options_parse(
     int argc, char* argv[], Plugin* plugins[], size_t nr_plugins, Error* error);