+++ src/plugin/Plugin.c
@@ -4,7 +4,7 @@
 #include "Plugin.h"
 
 
-void Resource_delete(Resource resource) {
+void Resource_delete(Resource* resource) {
     if (resource != NULL) {
         memset(resource, 0, sizeof(*resource));
         free(resource);
@@ -12,8 +12,8 @@ void Resource_delete(Resource resource) {
 }
 
 
-Resource Resource_new(char* name, int fd, Error* error) {
-    Resource resource = (Resource) malloc(sizeof(*resource));
+Resource* Resource_new(char* name, int fd, Error* error) {
+    Resource* resource = (Resource*) malloc(sizeof(*resource));
 
     if (resource == NULL) {
         Error_errno(error, errno);